Add examples of lists for the styleguide
Copy the content of 13-styleguide.html into 26-styleguide.html

Example of unordered list:
	inside main after Paragraph section, add :
	a new line and a comment with text Lists
	after, create a new section with inside:
	create a header with inside a level 2 heading with the text Lists
	after the new header, create a div with inside:
	a level 3 heading with text Unordered
	under it, add an unordered list with these items: Dolor pulvinar etiam magna etiam., Sagittis adipiscing lorem eleifend., Felis enim feugiat dolore viverra.

Example of ordered list:
	after previous unordered list, in the same div
	add a level 3 heading with text Ordered
	add an ordered list with these items:
	Dolor pulvinar etiam magna etiam.
	Sagittis adipiscing lorem eleifend.
	Felis enim feugiat dolore viverra.

Example of definition list:
after previous ordered list, in the same div
add a heading level 3 with text Definition
add a definition list with these items:
Term: Definition List title, Definition: Definition text.
Term: Startup, Definition: A startup company or startup is a company or temporary organization designed to search for a repeatable and scalable business model.
Term: Water, Definition: A colorless, transparent, odorless liquid that forms the seas, lakes, rivers, and rain and is the basis of the fluids of living organisms.