Fix layout issues
To easily view the website on a device of any screen/window size add
some media-queries for mobile and tablet.
Note: We already have a desktop version, we exceptionally add
this media-queries for mobile.

  For extra large devices (no media queries)
  For desktop / large devices (max-width: 992px)
  For tablet / medium styles (max-width: 767px)
  For mobile styles (max-width: 480px)

the media queries are added at the end of each section to facilitate the
reading but for performance reasons, the best practice is to unify all
similar breakpoints at the end of the CSS file.

The navbar is not allowing the website to fit the window. It's
temporarily hide and will create a mobile navbar later.