Skip links in 01-index.html
Add the <!-- Skip links --> comment

Create a new <nav> tag with the aria-label attribute.
Put the value Skip links inside it.

	Create a non ordered list of class off-screen
		Create a first li with a link inside
			Href: #a11y-primary-nav
			Class: skip-link
			Text: Skip to primary navigation
		Create a second li with a link inside
			Href: #a11y-main-content
			Class: skip-link
			Text: Skip to main content

On the <nav class="navbar-menu">, add an id with the text:
a11y-primary-nav and a tabindex="-1"

On the <main> tag, add an id with the text: a11y-main-content,
and tabindex="-1"