Change DOm with clousure
Task 5. Changing DOM with closure

- Function changeMode:
It accepts 5 arguments size (number), weight (string), transform
(string), background (string), color (string)
Using closure, it changes the style of the entire page and set the
font-size, font-weight, text-transform, background-color, and color

- Function main:
Set a variable named spooky that passes the arguments
9, bold, uppercase, pink, green to changeMode
Set a variable named darkMode that passes the arguments
12, bold, capitalize, black, white to changeMode
Set a variable named screamMode that passes the arguments
12, normal, lowercase, white, black to changeMode
Add a paragraph to the body of the page with the text Welcome Holberton!
Add a button to the body with the text Spooky
Add a button to the body with the text Dark mode
Add a button to the body with the text Scream mode
When clicking on each button, the page CSS change to the different
themes you created previously

- Requirements:
Use vanilla JavaScript to add the elements to the DOM and change the
style of the page
JQuery not allowed!