Measure the time needed to execute countPrimeNumbers
Task 9. Prime numbers & timing execution

Log to the console the time in milliseconds to execute the function
in this format:
Execution time of printing countPrimeNumbers was <time used> milliseconds.

Requirements:
Must use performance API to measure the time needed to execute the
function (performance.now())
Your code should display something in the range of Execution time of
printing countPrimeNumbers was 0.2700000002514571 milliseconds.